requirements.txt
flask
werkzeug
PyJWT
flask-login
flask-sqlalchemy
sqlalchemy
marshmallow